+++ include/GameState.hpp
@@ -45,13 +45,7 @@ public:
 
     @param newState The state to set the game's current state to.
    */
-  void setState(State newState);
-  /*
-    Set the current state to the state specified.
-
-    @param newState The state to set the game's current state to.
-   */
-  void setState(int newState);
+  void setState(State& newState);
 
 
   /*