+++ src/GameState.cpp
@@ -10,14 +10,11 @@
 
 GameState::GameState(shared_ptr<EventManager> eventManager){
   //potentially make an Init state in which the game initializes itself?
-  setState(State::MainMenu);
+  State main = State::MainMenu;
+  setState(main);
   this -> eventManager = eventManager;
 }
 
-void GameState::setState(State newState){
+void GameState::setState(State& newState){
   this->currentState=newState;
-}
-
-void GameState::setState(int newState){
-  this->currentState=(State)newState;
 }